.claude/worktrees/pensive-saha/requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-wtf==1.2.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.1
flask-limiter==3.5.0
flask-talisman==1.1.0
flask-caching==2.1.0
wtforms==3.1.2
email-validator==2.2.0
python-dotenv==1.0.1
werkzeug==3.0.1
#RPi.GPIO==0.7.1
w1thermsensor==2.0.0
apscheduler==3.10.4
numpy==1.26.4
smbus2==0.4.3
Pillow==10.4.0
